+++ configs.py
@@ -45,10 +45,14 @@ args.seq_len_ques       = 40
 args.seq_len_para       = 40
 args.seq_len_ans        = 12
 args.n_paras            = 40
-args.dim_hid            = 256
+args.d_embd             = 200
+args.d_hid              = 256
 args.max_len_ans        = 12
 args.teacher_forcing    = 0.3
 args.min_count_PGD      = 10        # min occurences of word to be added to vocab of PointerGeneratorDecoder
+args.d_vocab            = 124555    # Vocab size
+args.dropout            = 0.2
+args.n_layers           = 5
 
 
 ###############################